+++ CHANGELOG.md
@@ -14,6 +14,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
  - `StoreOperationDecorator` no longer raises `ValueError` when input of `operation` is not valid, 
 now it is logged as warning and record is ignored
+ - `Profiler.delete_metric` and `Profiler.delete_handler`no longer raising `KeyError`when metric/handler not exists
+now it is logged as warning 
 
 ## [0.2.2] - 2023-08-18